Upgraded ComonentRevisionsInfo to Angular.
Renamed component to ComponentStateInfoComponent.
Moved revisions dialog handling to WorkgroupComponentGrading.

#227
