+++ src/assets/wise5/classroomMonitor/classroomMonitorComponents/shared/shared.ts
@@ -1,7 +1,6 @@
 'use strict';
 
 import { ComponentNewWorkBadgeComponent } from '../../../../../app/classroom-monitor/component-new-work-badge/component-new-work-badge.component';
-import ComponentRevisionsInfo from './componentRevisionsInfo/componentRevisionsInfo';
 import MainMenu from './mainMenu/mainMenu';
 import NodeInfo from './nodeInfo/nodeInfo';
 import NotificationsMenu from './notificationsMenu/notificationsMenu';
@@ -30,7 +29,6 @@ const Shared = angular
     'componentNewWorkBadge',
     downgradeComponent({ component: ComponentNewWorkBadgeComponent }) as angular.IDirectiveFactory
   )
-  .component('componentRevisionsInfo', ComponentRevisionsInfo)
   .component('cmMainMenu', MainMenu)
   .component('notificationsMenu', NotificationsMenu)
   .component('nodeInfo', NodeInfo)
